refactor: make code provenance internal
Code provenance was implemented as part of the profiler product.
However, the feature is potentially useful to other products, so
we move it into the internal area to make it shareable.
