+++ tests/profiling/exporter/test_pprof.py
@@ -8,7 +8,6 @@ from ddtrace import ext
 from ddtrace.profiling.collector import _lock
 from ddtrace.profiling.collector import memalloc
 from ddtrace.profiling.collector import stack_event
-from ddtrace.profiling.exporter import _packages
 from ddtrace.profiling.exporter import pprof
 
 
@@ -701,7 +700,6 @@ def test_pprof_exporter(gan):
 
 @mock.patch("ddtrace.internal.utils.config.get_application_name")
 def test_pprof_exporter_libs(gan):
-    _packages._FILE_PACKAGE_MAPPING = _packages._build_package_file_mapping()
     gan.return_value = "bonjour"
     exp = pprof.PprofExporter()
     TEST_EVENTS = {
